// src/modules/payments/common/notification-event-types.enum.ts
export enum NotificationEventTypes {
  CANCEL = 'CANCEL', // +
  REFUND = 'REFUND', // +
  DID_RECOVER = 'DID_RECOVER', // +
  DID_RENEW = 'DID_RENEW', // +
  INTERACTIVE_RENEWAL = 'INTERACTIVE_RENEWAL', // +
  DID_CHANGE_RENEWAL_STATUS = 'DID_CHANGE_RENEWAL_STATUS',
  DID_FAIL_TO_RENEW = 'DID_FAIL_TO_RENEW',
  INITIAL_BUY = 'INITIAL_BUY',
  DID_CHANGE_RENEWAL_PREF = 'DID_CHANGE_RENEWAL_PREF',
  PRICE_INCREASE_CONSENT = 'PRICE_INCREASE_CONSENT',
}